fix(rust): resolve all clippy warnings, module inception and unnecessary clone/casts in workspace crates

+++ src-tauri/crates/media/src/utils/time.rs
@@ -63,7 +63,7 @@ fn civil_from_days(z: u64) -> (u64, u64, u64) {
     let d = doy - (153 * mp + 2) / 5 + 1; // [1, 31]
     let m = if mp < 10 { mp + 3 } else { mp - 9 }; // [1, 12]
     let y = if m <= 2 { y + 1 } else { y };
-    (y as u64, m as u64, d as u64)
+    (y as u64, m, d)
 }
 
 /// Format seconds as HH:MM:SS.mmm (dot separator, used by FFmpeg).
